requirements.txt
torch>=2.1
numpy>=1.24
matplotlib>=3.7
pandas>=2.0
PyYAML>=6.0
tqdm>=4.66
scikit-learn>=1.3
